css-utilities: refactor: split up functions from css-utilities, update test structure
