+++ packages/css-utilities/__tests__/fixtures/src/u-test-three.common.inline.scss
@@ -1,0 +1,3 @@
+.u-test-three {
+	display: flex;
+}
