+++ packages/css-utilities/lib/utils/utils.js
@@ -1,0 +1,5 @@
+const getScssPathsWithExtension = require( './getScssPathsWithExtension' );
+
+module.exports = {
+	getScssPathsWithExtension: getScssPathsWithExtension
+}
